+++ CHANGELOG.md
@@ -65,6 +65,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Show progress when loading variant categories (#97)
 - Variant loading filter based on rank score and category (#98)
 - Parse custom images for STR variants (#102)
+- Support loading VCFs by category and research status (#111)
 ### Changed
 - Structure of parsed entities, to reflect documents saved in the database (#20)
 - Improved and simplified code (#22)